Fixed compilation issue

+++ source/lib/ffglquickstart/FFGLAudio.cpp
@@ -18,7 +18,7 @@ void Audio::Update( std::vector< float > _fft )
 	float currentVolStore = 0;
 	float bassStore       = 0;
 	float bassCount       = 0;
-	int bassSep           = fft.size() / 3;
+	int bassSep           = static_cast< int >( fft.size() / 3 );
 	float medStore        = 0;
 	float medCount        = 0;
 	int highSep           = bassSep * 2;
